Add new functions that allow rendering a single and or all projects in appState
Added function renderProject(project), which renders a single project
object to the DOM. Added a container 'main-content-container' that
contains 'project-cotainer' and a new 'project-task-container' that will
contain tasks associated with a project. Added getProjectsFromStorage()
which gets all projects from localStorage and returns an array of project
objects. Added loadsProjectIntoAppState() that loads all projects into
appState.projects. Add appState that contains an empty array that will be
populated in index.js with loadProjectsIntoAppState().